+++ include/erl_common/block_timer.hpp
@@ -12,11 +12,13 @@ namespace erl::common {
     struct BlockTimer {
         const char *label;
         double *dt;
+        bool verbose;
         std::chrono::time_point<std::chrono::high_resolution_clock> t1;
 
-        explicit BlockTimer(const char *label, double *dt = nullptr)
+        explicit BlockTimer(const char *label, double *dt = nullptr, const bool verbose = true)
             : label(label),
               dt(dt),
+              verbose(verbose),
               t1(std::chrono::high_resolution_clock::now()) {}
 
         ~BlockTimer() {
@@ -24,24 +26,26 @@ namespace erl::common {
             double &&dt = std::chrono::duration<double, typename Duration::period>(t2 - t1).count();
             if (this->dt != nullptr) { *this->dt = dt; }
 
-            std::string unit;
-            if (std::is_same_v<Duration, std::chrono::nanoseconds>) {
-                unit = " ns";
-            } else if (std::is_same_v<Duration, std::chrono::microseconds>) {
-                unit = " us";
-            } else if (std::is_same_v<Duration, std::chrono::milliseconds>) {
-                unit = " ms";
-            } else if (std::is_same_v<Duration, std::chrono::seconds>) {
-                unit = " s";
-            } else if (std::is_same_v<Duration, std::chrono::minutes>) {
-                unit = " min";
-            } else if (std::is_same_v<Duration, std::chrono::hours>) {
-                unit = " hrs";
-            }
+            if (verbose) {
+                std::string unit;
+                if (std::is_same_v<Duration, std::chrono::nanoseconds>) {
+                    unit = " ns";
+                } else if (std::is_same_v<Duration, std::chrono::microseconds>) {
+                    unit = " us";
+                } else if (std::is_same_v<Duration, std::chrono::milliseconds>) {
+                    unit = " ms";
+                } else if (std::is_same_v<Duration, std::chrono::seconds>) {
+                    unit = " s";
+                } else if (std::is_same_v<Duration, std::chrono::minutes>) {
+                    unit = " min";
+                } else if (std::is_same_v<Duration, std::chrono::hours>) {
+                    unit = " hrs";
+                }
 
-            std::string msg = fmt::format("{}: {:.3f}{}", label, dt, unit);
-            if (ProgressBar::GetNumBars() == 0) { msg += "\n"; }
-            ProgressBar::Write(msg);
+                std::string msg = fmt::format("{}: {:.3f}{}", label, dt, unit);
+                if (ProgressBar::GetNumBars() == 0) { msg += "\n"; }
+                ProgressBar::Write(msg);
+            }
         }
     };
 }  // namespace erl::common