+++ src/pangolin_plotter_curve_2d.cpp
@@ -1,4 +1,4 @@
-#include "erl_common/pangolin_plotter_time_series_2d.hpp"
+#include "erl_common/pangolin_plotter_curve_2d.hpp"
 
 #include "erl_common/logging.hpp"
 
@@ -6,19 +6,19 @@
 
 namespace erl::common {
 
-    PangolinPlotterTimeSeries2D::PangolinPlotterTimeSeries2D(
+    PangolinPlotterCurve2D::PangolinPlotterCurve2D(
         std::shared_ptr<PangolinWindow> window,
         const std::string &plotter_name,
         const std::vector<std::string> &labels,
         const int plot_window_width,
-        const float plot_t_init,
-        const float plot_t_tick,
+        const float plot_x_init,
+        const float plot_x_tick,
         const float plot_y_tick,
         const pangolin::Colour &bg_color,
         const pangolin::Colour &axis_color,
         const pangolin::Colour &tick_color)
         : m_window_(std::move(window)),
-          m_plotter_(&m_data_log_, plot_t_init, plot_t_init + static_cast<float>(plot_window_width) * plot_t_tick, -1.0f, 1.0f, plot_t_tick, plot_y_tick),
+          m_plotter_(&m_data_log_, plot_x_init, plot_x_init + static_cast<float>(plot_window_width) * plot_x_tick, -1.0f, 1.0f, plot_x_tick, plot_y_tick),
           m_display_(m_window_->AddDisplay(plotter_name, m_plotter_)),
           m_log_buffer_(labels.size()),
           m_max_values_(plot_window_width, std::numeric_limits<float>::lowest()),
@@ -34,7 +34,7 @@ namespace erl::common {
     }
 
     void
-    PangolinPlotterTimeSeries2D::Append(const float t, const std::vector<float> &values) {
+    PangolinPlotterCurve2D::Append(const float x, const std::vector<float> &values) {
         ERL_ASSERTM(values.size() == m_log_buffer_.size() - 1, "Number of values does not match the number of series in the plotter");
 
         m_window_->Activate();
@@ -43,13 +43,13 @@ namespace erl::common {
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
         // log data
-        m_log_buffer_[0] = t;
+        m_log_buffer_[0] = x;
         std::copy(values.begin(), values.end(), m_log_buffer_.begin() + 1);
         m_data_log_.Log(m_log_buffer_);
 
         // scroll along x-axis
         auto range = m_plotter_.GetView();
-        m_plotter_.ScrollViewSmooth(t > range.x.max ? t - range.x.max : 0.0f, 0.0f);
+        m_plotter_.ScrollViewSmooth(x > range.x.max ? x - range.x.max : 0.0f, 0.0f);
         const auto [min, max] = std::minmax_element(values.begin(), values.end());
         m_min_values_[m_head_] = *min;
         m_max_values_[m_head_++] = *max;